pyproject.toml
[project]
name = "scoped-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "aiofiles==25.1.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.2",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "attrs==25.4.0",
    "av>=16.1.0",
    "backoff==2.2.1",
    "beautifulsoup4==4.14.2",
    "certifi==2025.11.12",
    "cffi==2.0.0",
    "charset-normalizer==3.4.4",
    "chromadb>=1.4.1",
    "click==8.3.1",
    "cryptography==46.0.3",
    "ctranslate2>=4.6.3",
    "dataclasses-json==0.6.7",
    "docling>=2.68.0",
    "docling-core>=2.59.0",
    "easyocr>=1.7.2",
    "emoji==2.15.0",
    "faiss-cpu==1.13.0",
    "fastapi[standard]>=0.100.0",
    "faster-whisper>=1.2.1",
    "filelock==3.20.0",
    "filetype==1.2.0",
    "frozenlist==1.8.0",
    "fsspec==2025.10.0",
    "google-generativeai>=0.8.6",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "html5lib==1.1",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "huggingface-hub==0.36.0",
    "idna==3.11",
    "imageio-ffmpeg>=0.6.0",
    "joblib==1.5.2",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-classic==1.0.0",
    "langchain-community==0.4.1",
    "langchain-core==1.1.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface==1.1.0",
    "langchain-ollama>=1.0.0",
    "langchain-text-splitters==1.0.0",
    "langdetect==1.0.9",
    "langgraph==1.0.4",
    "langgraph-checkpoint==3.0.1",
    "langgraph-prebuilt==1.0.5",
    "langgraph-sdk==0.2.10",
    "langsmith==0.4.49",
    "lxml==6.0.2",
    "marshmallow==3.26.1",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "nltk==3.9.2",
    "numpy==2.3.5",
    "olefile==0.47",
    "ollama>=0.3.0",
    "openai>=2.15.0",
    "openpyxl>=3.1.5",
    "orjson==3.11.4",
    "ormsgpack==1.12.0",
    "packaging==25.0",
    "pillow>=11.3.0",
    "propcache==0.4.1",
    "psutil==7.1.3",
    "pycparser==2.23",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pypdf==6.4.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv==1.2.1",
    "python-iso639==2025.11.16",
    "python-magic==0.4.27",
    "python-multipart>=0.0.20",
    "python-oxmsg==0.0.2",
    "python-pptx>=1.0.2",
    "pyyaml==6.0.3",
    "rank-bm25>=0.2.2",
    "rapidfuzz==3.14.3",
    "regex==2025.11.3",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "sentence-transformers>=5.1.2",
    "six==1.17.0",
    "soupsieve==2.8",
    "sqlalchemy==2.0.44",
    "streamlit>=1.28.0",
    "tenacity==9.1.2",
    "tokenizers==0.22.1",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "tqdm==4.67.1",
    "transformers>=4.57.3",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "unstructured==0.18.21",
    "unstructured-client==0.42.4",
    "urllib3==2.5.0",
    "uvicorn>=0.23.0",
    "webencodings==0.5.1",
    "wrapt==2.0.1",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
